QT: Create inventory instead of using cursor stack
Before, we would use the cursor stack to import the default template.
However, this could cause issues if the scenario happens to put
something in the cursor right at the beginning. To avoid this, we now
use the brand-new script inventories feature to create our own
inventory for the blueprint.